chore: fix directory tree
