+++ mod.ts
@@ -1,16 +1,1 @@
-export { getToken } from "./token.ts";
-export {
-  getNoticesFromKupid,
-  getSchedulesFromKupid,
-  getScholarFromKupid,
-  parseNoticeInfo,
-  parseScheduleInfo,
-  parseScholarInfo,
-} from "./kupid.ts";
-export {
-  type KupidType,
-  type NoticeInfo,
-  fetchKupidNotices,
-  fetchParsedKupidNoptices,
-} from "./wrapper.ts";
-export { type InfoCollegeType, getNoticeInfos } from "./info.ts";
+export * from "./src/index.ts";
