+++ src/kupid/kupid.ts
@@ -1,6 +1,5 @@
-import { getToken } from "./token.ts";
-import { NoticeInfo } from "./wrapper.ts";
-import { trim } from "./utils.ts";
+import type { NoticeInfo, KupidType } from "../type.d.ts";
+import { getToken, trim } from "../utils.ts";
 
 interface GrwResponse {
   grwSessionId: string;
@@ -121,7 +120,7 @@ function makeFilePathPublic(html: string) {
  * @param password KUPID password
  * @returns KUPID '일반공지' 공지사항 10개의 HTML string Array를 반환합니다.
  */
-export async function getNoticesFromKupid(id: string, password: string): Promise<string[]> {
+async function getNoticesFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "11");
   const params = parseNoticeParamsFromHTML(html);
@@ -140,7 +139,7 @@ export async function getNoticesFromKupid(id: string, password: string): Promise
  * @param password KUPID password
  * @returns KUPID '학사일정' 공지사항 10개의 HTML string Array를 반환합니다.
  */
-export async function getSchedulesFromKupid(id: string, password: string): Promise<string[]> {
+async function getSchedulesFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "89");
   const params = parseScheduleParamsFromHTML(html);
@@ -158,7 +157,7 @@ export async function getSchedulesFromKupid(id: string, password: string): Promi
  * @param password KUPID password
  * @returns KUPID '학사일정' 공지사항 10개의 HTML string Array를 반환합니다.
  */
-export async function getScholarFromKupid(id: string, password: string): Promise<string[]> {
+async function getScholarFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "88");
   const params = parseNoticeParamsFromHTML(html);
@@ -175,7 +174,7 @@ export async function getScholarFromKupid(id: string, password: string): Promise
  * @param html getNoticesFromKupid 함수에서 반환된 HTML string
  * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
  */
-export function parseNoticeInfo(html: string): NoticeInfo {
+function parseNoticeInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
   const rawWriter = tableRows[0].match(/\<td\>(.+)\<\/td\>/);
@@ -209,7 +208,7 @@ export function parseNoticeInfo(html: string): NoticeInfo {
  * @param html getSchedulesFromKupid 함수에서 반환된 HTML string
  * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
  */
-export function parseScheduleInfo(html: string): NoticeInfo {
+function parseScheduleInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
   const rawWriter = tableRows[0].match(/\<td\>(.+)\<\/td\>/);
@@ -244,7 +243,7 @@ export function parseScheduleInfo(html: string): NoticeInfo {
  * @param html getScholarFromKupid 함수에서 반환된 HTML string
  * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
  */
-export function parseScholarInfo(html: string): NoticeInfo {
+function parseScholarInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
   const rawWriter = tableRows[0].match(/\<td\>(.+)\<\/td\>/);
@@ -273,4 +272,53 @@ export function parseScholarInfo(html: string): NoticeInfo {
     content,
     url,
   };
+}
+
+/**
+ *
+ * @param id KUPID id
+ * @param password KUPID password
+ * @param type 장학공지 Scholar, 일반공지 Notice, 학사일정 Schedule
+ * @returns KUPID의 해당 카테고리 공지사항 최근 10개의 raw HTML string array를 반환합니다.
+ */
+export async function fetchKupidNotices(
+  id: string,
+  password: string,
+  type: KupidType
+): Promise<string[]> {
+  switch (type) {
+    case "Scholar":
+      return getScholarFromKupid(id, password);
+    case "Notice":
+      return getNoticesFromKupid(id, password);
+    case "Schedule":
+      return getSchedulesFromKupid(id, password);
+    default:
+      throw new Error("Invalid Type");
+  }
+}
+
+/**
+ *
+ * @param id KUPID id
+ * @param password KUPID password
+ * @param type 장학공지 Scholar, 일반공지 Notice, 학사일정 Schedule
+ * @returns KUPID의 해당 카테고리 공지사항 최근 10개의 content HTML, 제목, 작성자, 게시일자, KUPID 내부 id, public URL을 반환합니다.
+ */
+export async function fetchParsedKupidNotices(
+  id: string,
+  password: string,
+  type: KupidType
+): Promise<NoticeInfo[]> {
+  const htmls = await fetchKupidNotices(id, password, type);
+  switch (type) {
+    case "Scholar":
+      return htmls.map(parseScholarInfo);
+    case "Notice":
+      return htmls.map(parseNoticeInfo);
+    case "Schedule":
+      return htmls.map(parseScheduleInfo);
+    default:
+      throw new Error("Invalid Type");
+  }
 }
