+++ src/type-a/ee.ts
