+++ src/type-a/type-a.ts
