+++ src/type.d.ts
@@ -1,0 +1,10 @@
+export interface NoticeInfo {
+  id: string;
+  title: string;
+  date: string;
+  writer: string;
+  content: string;
+  url: string;
+}
+
+export type KupidType = "Scholar" | "Notice" | "Schedule";
