--- utils.ts
@@ -1,3 +1,0 @@
-export function trim(str: string): string {
-  return str.replace(/\<(.+)\>/g, "").replace(/(&nbsp;)+/g, " ");
-}
